+++ api/handlers.go
@@ -50,7 +50,7 @@ func getInfo(c *context, w http.ResponseWriter, r *http.Request) {
 		ServerVersion:     "swarm/" + version.VERSION,
 		OperatingSystem:   runtime.GOOS,
 		Architecture:      runtime.GOARCH,
-		NCPU:              c.cluster.TotalCpus(),
+		NCPU:              int(c.cluster.TotalCpus()),
 		MemTotal:          c.cluster.TotalMemory(),
 		HTTPProxy:         os.Getenv("http_proxy"),
 		HTTPSProxy:        os.Getenv("https_proxy"),
@@ -401,7 +401,20 @@ func getContainersJSON(c *context, w http.ResponseWriter, r *http.Request) {
 		if !filters.Match("node", container.Engine.Name) {
 			continue
 		}
-
+		if filters.Include("volume") {
+			volumeExist := fmt.Errorf("volume mounted in container")
+			err := filters.WalkValues("volume", func(value string) error {
+				for _, mount := range container.Info.Mounts {
+					if mount.Name == value || mount.Destination == value {
+						return volumeExist
+					}
+				}
+				return nil
+			})
+			if err != volumeExist {
+				continue
+			}
+		}
 		if len(filtExited) > 0 {
 			shouldSkip := true
 			for _, code := range filtExited {
@@ -608,7 +621,7 @@ func deleteContainers(c *context, w http.ResponseWriter, r *http.Request) {
 
 // POST /networks/create
 func postNetworksCreate(c *context, w http.ResponseWriter, r *http.Request) {
-	var request apitypes.NetworkCreate
+	var request apitypes.NetworkCreateRequest
 
 	if err := json.NewDecoder(r.Body).Decode(&request); err != nil {
 		httpError(w, err.Error(), http.StatusBadRequest)
@@ -619,7 +632,7 @@ func postNetworksCreate(c *context, w http.ResponseWriter, r *http.Request) {
 		request.Driver = "overlay"
 	}
 
-	response, err := c.cluster.CreateNetwork(&request)
+	response, err := c.cluster.CreateNetwork(request.Name, &request.NetworkCreate)
 	if err != nil {
 		httpError(w, err.Error(), http.StatusInternalServerError)
 		return
@@ -798,7 +811,7 @@ func postContainersStart(c *context, w http.ResponseWriter, r *http.Request) {
 	}
 	r.Body.Close()
 
-	if len(buf) <= 2 {
+	if len(buf) <= 2 || (len(buf) == 4 && string(buf) == "null") {
 		hostConfig = nil
 	} else {
 		if err := json.Unmarshal(buf, hostConfig); err != nil {
@@ -1054,6 +1067,7 @@ func proxyContainer(c *context, w http.ResponseWriter, r *http.Request) {
 	if err != nil {
 		if container == nil {
 			httpError(w, err.Error(), http.StatusNotFound)
+			return
 		}
 		httpError(w, err.Error(), http.StatusInternalServerError)
 		return
@@ -1169,8 +1183,10 @@ func postTagImage(c *context, w http.ResponseWriter, r *http.Request) {
 	if err := c.cluster.TagImage(name, repo, tag, force); err != nil {
 		if strings.HasPrefix(err.Error(), "No such image") {
 			httpError(w, err.Error(), http.StatusNotFound)
+			return
 		} else {
 			httpError(w, err.Error(), http.StatusInternalServerError)
+			return
 		}
 	}
 	w.WriteHeader(http.StatusCreated)
@@ -1212,6 +1228,7 @@ func postCommit(c *context, w http.ResponseWriter, r *http.Request) {
 	if err != nil {
 		if container == nil {
 			httpError(w, err.Error(), http.StatusNotFound)
+			return
 		}
 		httpError(w, err.Error(), http.StatusInternalServerError)
 		return
@@ -1261,7 +1278,6 @@ func postBuild(c *context, w http.ResponseWriter, r *http.Request) {
 		CPUSetMems:     r.Form.Get("cpusetmems"),
 		CgroupParent:   r.Form.Get("cgroupparent"),
 		ShmSize:        int64ValueOrZero(r, "shmsize"),
-		Context:        r.Body,
 	}
 
 	buildArgsJSON := r.Form.Get("buildargs")
@@ -1290,7 +1306,7 @@ func postBuild(c *context, w http.ResponseWriter, r *http.Request) {
 	w.Header().Set("Content-Type", "application/json")
 	wf := NewWriteFlusher(w)
 
-	err := c.cluster.BuildImage(buildImage, wf)
+	err := c.cluster.BuildImage(r.Body, buildImage, wf)
 	if err != nil {
 		httpError(w, err.Error(), http.StatusInternalServerError)
 	}
@@ -1302,6 +1318,7 @@ func postRenameContainer(c *context, w http.ResponseWriter, r *http.Request) {
 	if err != nil {
 		if container == nil {
 			httpError(w, err.Error(), http.StatusNotFound)
+			return
 		}
 		httpError(w, err.Error(), http.StatusInternalServerError)
 		return
@@ -1315,8 +1332,10 @@ func postRenameContainer(c *context, w http.ResponseWriter, r *http.Request) {
 	if err = c.cluster.RenameContainer(container, r.Form.Get("name")); err != nil {
 		if strings.HasPrefix(err.Error(), "Conflict") {
 			httpError(w, err.Error(), http.StatusConflict)
+			return
 		} else {
 			httpError(w, err.Error(), http.StatusInternalServerError)
+			return
 		}
 	}
 	w.WriteHeader(http.StatusNoContent)
@@ -1328,6 +1347,7 @@ func proxyHijack(c *context, w http.ResponseWriter, r *http.Request) {
 	if err != nil {
 		if container == nil {
 			httpError(w, err.Error(), http.StatusNotFound)
+			return
 		}
 		httpError(w, err.Error(), http.StatusInternalServerError)
 		return